Move functions to the data_utils
